G3-26 Transaction List Fixed
Fixed Transaction List
